+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## [master](https://github.com/anmonteiro/lumo/compare/1.0.0-alpha2...HEAD) (unreleased)
+## [master](https://github.com/anmonteiro/lumo/compare/1.0.0...HEAD) (unreleased)
+
+### Bug fixes
+
+- Regression where specifying a main script to be executed would print the banner
+as if entering the REPL.
+
+## [1.0.0](https://github.com/anmonteiro/lumo/compare/1.0.0-alpha3...1.0.0) (2016-11-22)
 
 ### New features
 
@@ -21,6 +28,7 @@ at runtime ([#28](https://github.com/anmonteiro/lumo/issues/28)).
 the order they appear. Additionally, similarly to the Clojure REPL, arguments that
 appear after a main option will be bound to the `lumo.core/*command-line-args*`
 var, a seq of strings.
+- Print a newline when pressing ctrl+C or evaluating empty form ([#29](https://github.com/anmonteiro/lumo/issues/29)).
 
 ## [1.0.0-alpha3](https://github.com/anmonteiro/lumo/compare/1.0.0-alpha2...1.0.0-alpha3) (2016-11-15)